fileservice: remove CacheFS; add memory cache to S3FS (#4365)
fileservice: remove CacheFS; add memory cache to S3FS

Approved by: @lni
