+++ pkg/fileservice/local_etl_fs_test.go
@@ -31,13 +31,4 @@ func TestLocalETLFS(t *testing.T) {
 		})
 	})
 
-	t.Run("cache", func(t *testing.T) {
-		testCache(t, func() FileService {
-			dir := t.TempDir()
-			fs, err := NewLocalETLFS(dir)
-			assert.Nil(t, err)
-			return fs
-		})
-	})
-
 }
