+++ pkg/fileservice/local_fs_test.go
@@ -49,15 +49,6 @@ func TestLocalFS(t *testing.T) {
 		})
 	})
 
-	t.Run("cache", func(t *testing.T) {
-		testCache(t, func() FileService {
-			dir := t.TempDir()
-			fs, err := NewLocalFS(dir)
-			assert.Nil(t, err)
-			return fs
-		})
-	})
-
 }
 
 func BenchmarkLocalFS(b *testing.B) {
