+++ pkg/fileservice/memory_fs_test.go
@@ -38,14 +38,6 @@ func TestMemoryFS(t *testing.T) {
 		})
 	})
 
-	t.Run("cache", func(t *testing.T) {
-		testCache(t, func() FileService {
-			fs, err := NewMemoryFS()
-			assert.Nil(t, err)
-			return fs
-		})
-	})
-
 }
 
 func BenchmarkMemoryFS(b *testing.B) {
